% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{house_reps_regular}
\alias{house_reps_regular}
\title{Oregon House Representatives in regular sessions}
\format{A tibble with 7381 rows and 13 variables where each row represents 
a legislator who was in office at the start of a particular legislative session:
\describe{
  \item{session_name}{Legislative session}
  \item{legislator}{Name of representative}
  \item{district}{The legislative district they represented in this session}
  \item{party}{The affiliation of the legislator at the start of the session.}
}}
\source{
\url{https://www.oregonlegislature.gov/chief-clerk/Pages/representatives.aspx}
}
\usage{
house_reps_regular
}
\description{
A dataset containing the chronological list of Oregon House representatives 
in regular sessions since Oregon was a state (1860) to present.
}
\examples{
house_reps_regular
}
\keyword{datasets}
