% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{representatives}
\alias{representatives}
\title{Oregon House Representatives}
\format{A tibble with 7381 rows and 13 variables where each row represents 
a legislator in a particular legislative session:
\describe{
  \item{session_name}{Legislative session}
  \item{Legislator}{Name of representative}
  \item{Distric}{The legislative district they represented in this session}
  ...
}}
\source{
\url{https://www.oregonlegislature.gov/chief-clerk/Pages/representatives.aspx}
}
\usage{
representatives
}
\description{
A dataset containing the chronological list of Oregon House representatives 
since 1849 to present.
}
\keyword{datasets}
